CPM-550: Move ProductWasCreated and ProductWasUpdated in API
