+++ src/Akeneo/Pim/Enrichment/Product/back/API/Event/ProductWasUpdated.php
@@ -2,13 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Akeneo\Pim\Enrichment\Product\Domain\Event;
+namespace Akeneo\Pim\Enrichment\Product\API\Event;
 
 /**
  * @copyright 2022 Akeneo SAS (https://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ProductWasCreated
+class ProductWasUpdated
 {
     public function __construct(private string $productIdentifier)
     {
